+++ app/Http/Controllers/Events/EventICalExport.php
@@ -71,14 +71,32 @@ class EventICalExport extends Controller
             /** @var $member Member */
 
             /** @var Carbon $birthday */
+            $birthdayLastYear = $member->birthday->clone();
+            $birthdayLastYear->setYear($currentYear-1);
+
+            //last year birthday
+            $calEvent = Event::create($member->getFullName());
+            $calEvent->startsAt($birthdayLastYear);
+            $calEvent->description((string)($currentYear - 1 - $member->birthday->year));
+            $calEvent->fullDay();
+            $calendar->event($calEvent);
+
+            //this year
             $birthday = $member->birthday->clone();
             $birthday->setYear($currentYear);
-
             $calEvent = Event::create($member->getFullName());
             $calEvent->startsAt($birthday);
             $calEvent->description((string)($currentYear - $member->birthday->year));
             $calEvent->fullDay();
+            $calendar->event($calEvent);
 
+            //next year
+            $birthdayNextYear = $member->birthday->clone();
+            $birthdayNextYear->setYear($currentYear+1);
+            $calEvent = Event::create($member->getFullName());
+            $calEvent->startsAt($birthdayNextYear);
+            $calEvent->description((string)($currentYear + 1 - $member->birthday->year));
+            $calEvent->fullDay();
             $calendar->event($calEvent);
         }
     }